Make RoslynToolPath a dynamic property so copy occurs on restore
RoslynToolPath (and subsequently the list of roslyn files to copy) was calculated once since it was done outside of a target. By moving it into a target, it will be recalculated on each invocation. This also updates the DependsOnTarget attributes to ensure calculation is done when needed.

This change also removes the install.ps1/uninstall.ps1 since those are no longer needed.